Add Apply Coupon button and action handler to CouponsView

+++ kioske.sql
@@ -81,5 +81,5 @@ INSERT INTO Product_Tags (product_id, tag_id) VALUES
 CREATE TABLE IF NOT EXISTS Coupons
 (
     code        VARCHAR(255)   NOT NULL,
-    discount    INT NOT NULL CHECK (discount >= 0)
+    discount    DOUBLE NOT NULL CHECK (discount >= 0)
 );
